chore(checkout): reorganize auth pages routing
